feat: create a global component
* generateAudioNumbers creates the right number audio and outputs the romanized text of the number
